update from the original chrono project

+++ src/chrono_opengl/ChOpenGLViewer.h
@@ -27,8 +27,9 @@
 #include "chrono_opengl/UI/ChOpenGLContacts.h"
 #include "chrono_opengl/UI/ChOpenGLHUD.h"
 #include "chrono_opengl/UI/ChOpenGLGraphs.h"
-#include "physics/ChSystem.h"
-#include "core/ChTimer.h"
+
+#include "chrono/physics/ChSystem.h"
+#include "chrono/core/ChTimer.h"
 
 //#include "chrono_parallel/physics/ChSystemParallel.h"
 #include <glfw3.h>
@@ -36,7 +37,7 @@
 namespace chrono {
 namespace opengl {
 
-/// @addtogroup opengl
+/// @addtogroup opengl_module
 /// @{
 
 enum RenderMode { POINTS, WIREFRAME, SOLID };
@@ -90,7 +91,7 @@ class CH_OPENGL_API ChOpenGLViewer : public ChOpenGLBase {
     ChOpenGLCloud fea_nodes;
     ChOpenGLWires fea_elements;
 
-	ChOpenGLWires line_path;
+    ChOpenGLWires line_path;
 
     ChOpenGLContacts contact_renderer;
     ChOpenGLHUD HUD_renderer;
@@ -103,11 +104,11 @@ class CH_OPENGL_API ChOpenGLViewer : public ChOpenGLBase {
     std::vector<glm::vec3> grid_data;
     std::vector<glm::vec3> mpm_grid_data;
     std::vector<glm::vec3> mpm_node_data;
-	std::vector<glm::vec3> line_path_data;
+    std::vector<glm::vec3> line_path_data;
 
-    int simulation_frame;   // The current frame number
-    float simulation_h;     // The simulation step size
-    float simulation_time;  // The current simulation time
+    int simulation_frame;    // The current frame number
+    double simulation_h;     // The simulation step size
+    double simulation_time;  // The current simulation time
     bool pause_sim;
     bool pause_vis;
     bool single_step;
@@ -121,8 +122,8 @@ class CH_OPENGL_API ChOpenGLViewer : public ChOpenGLBase {
 
     glm::mat4 model, view, projection, modelview;
 
-    float old_time, current_time, time_total, time_text, time_geometry;
-    float fps;
+    double old_time, current_time, time_total, time_text, time_geometry;
+    double fps;
 
     std::vector<glm::mat4> model_box;
     std::vector<glm::mat4> model_sphere;
@@ -132,6 +133,7 @@ class CH_OPENGL_API ChOpenGLViewer : public ChOpenGLBase {
     std::map<std::string, std::vector<glm::mat4> > model_obj;
 };
 
-/// @} opengl
+/// @} opengl_module
+
 }
 }
