+++ src/chrono_opengl/UI/ChOpenGLHUD.h
@@ -25,10 +25,11 @@
 
 #include "chrono/physics/ChSystem.h"
 //#include "chrono_parallel/physics/ChSystemParallel.h"
+
 namespace chrono {
 namespace opengl {
 
-/// @addtogroup opengl
+/// @addtogroup opengl_module
 /// @{
 
 /// Class that renders the text and other UI elements.
@@ -46,11 +47,11 @@ class CH_OPENGL_API ChOpenGLHUD : public ChOpenGLBase {
     void GenerateExtraStats(ChSystem* physics_system);
     void TakeDown();
     void Update(const glm::ivec2& window_size,
-                const float& dpi,
-                const float& frame_per_sec,
-                const float& t_geometry,
-                const float& t_text,
-                const float& t_total);
+                const double& dpi,
+                const double& frame_per_sec,
+                const double& t_geometry,
+                const double& t_text,
+                const double& t_total);
     void Draw();
 
   private:
@@ -63,14 +64,16 @@ class CH_OPENGL_API ChOpenGLHUD : public ChOpenGLBase {
     float aspect, z_x, z_y;
     char buffer[50];
     ChOpenGLCamera* render_camera;
-    float time_geometry, time_text, time_total, fps;
+    double time_geometry, time_text, time_total, fps;
 
     ChTimer<>* timer_render;
     ChTimer<>* timer_text;
     ChTimer<>* timer_geometry;
 };
 
-/// @} opengl
+/// @} opengl_module
+
 }
 }
-#endif  // END of CHOPENGLHUD_H
+
+#endif
